// app/apis/pantmig-api/apis/NotificationsApi.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PantmigService API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import * as runtime from '../runtime';
import type {
  MarkReadRequest,
} from '../models/index';
import {
    MarkReadRequestFromJSON,
    MarkReadRequestToJSON,
} from '../models/index';

export interface NotificationsMarkReadPostRequest {
    markReadRequest: MarkReadRequest;
}

export interface NotificationsRecentGetRequest {
    take: number;
}

/**
 * 
 */
export class NotificationsApi extends runtime.BaseAPI {

    /**
     */
    async notificationsMarkReadPostRaw(requestParameters: NotificationsMarkReadPostRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<void>> {
        if (requestParameters['markReadRequest'] == null) {
            throw new runtime.RequiredError(
                'markReadRequest',
                'Required parameter "markReadRequest" was null or undefined when calling notificationsMarkReadPost().'
            );
        }

        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        headerParameters['Content-Type'] = 'application/json';

        if (this.configuration && this.configuration.apiKey) {
            headerParameters["Authorization"] = await this.configuration.apiKey("Authorization"); // Bearer authentication
        }


        let urlPath = `/notifications/mark-read`;

        const response = await this.request({
            path: urlPath,
            method: 'POST',
            headers: headerParameters,
            query: queryParameters,
            body: MarkReadRequestToJSON(requestParameters['markReadRequest']),
        }, initOverrides);

        return new runtime.VoidApiResponse(response);
    }

    /**
     */
    async notificationsMarkReadPost(requestParameters: NotificationsMarkReadPostRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<void> {
        await this.notificationsMarkReadPostRaw(requestParameters, initOverrides);
    }

    /**
     */
    async notificationsRecentGetRaw(requestParameters: NotificationsRecentGetRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<void>> {
        if (requestParameters['take'] == null) {
            throw new runtime.RequiredError(
                'take',
                'Required parameter "take" was null or undefined when calling notificationsRecentGet().'
            );
        }

        const queryParameters: any = {};

        if (requestParameters['take'] != null) {
            queryParameters['take'] = requestParameters['take'];
        }

        const headerParameters: runtime.HTTPHeaders = {};

        if (this.configuration && this.configuration.apiKey) {
            headerParameters["Authorization"] = await this.configuration.apiKey("Authorization"); // Bearer authentication
        }


        let urlPath = `/notifications/recent`;

        const response = await this.request({
            path: urlPath,
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        }, initOverrides);

        return new runtime.VoidApiResponse(response);
    }

    /**
     */
    async notificationsRecentGet(requestParameters: NotificationsRecentGetRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<void> {
        await this.notificationsRecentGetRaw(requestParameters, initOverrides);
    }

}